+++ backend/canisters/group/impl/src/updates/update_group_v2.rs
@@ -91,13 +91,17 @@ fn prepare(args: &Args, state: &RuntimeState) -> Result<PrepareResult, Response>
     let gate = args.gate.as_ref().apply_to(state.data.chat.gate.value.as_ref());
 
     if let Some(member) = state.data.get_member(caller) {
+        let permissions_prev = args.permissions.clone().map(|ps| ps.into());
+        let permissions = args.permissions_v2.as_ref().or(permissions_prev.as_ref());
+
+        //.or(args.permissions.map(|ps| ps.into()));
         match state.data.chat.can_update(
             &member.user_id,
             &args.name,
             &args.description,
             &args.rules,
             &args.avatar,
-            &args.permissions,
+            permissions,
             &args.public,
         ) {
             Ok(_) => {
@@ -140,7 +144,7 @@ fn commit(my_user_id: UserId, args: Args, state: &mut RuntimeState) -> SuccessRe
         args.description,
         args.rules,
         args.avatar,
-        args.permissions,
+        args.permissions_v2.or(args.permissions.map(|ps| ps.into())),
         args.gate,
         args.public,
         args.events_ttl,